+++ Emulator/Cartridges/CustomCartridges/EasyFlash.h
@@ -92,7 +92,7 @@ private:
     // Handling ROM packets
     //
     
-    void loadChip(unsigned nr, const CRTFile &crt) override;
+    void loadChip(isize nr, const CRTFile &crt) override;
 
     
     //
